+++ easylogging++.h
@@ -1,7 +1,7 @@
 ///////////////////////////////////////////////////////////////////////////
 //                                                                       //
 // easylogging++.h - Core of EasyLogging++                               //
-//   EasyLogging++ v3.29                                                 //
+//   EasyLogging++ v3.xx                                                 //
 //   Cross platform logging made easy for C++ applications               //
 //   Author Majid Khan <mkhan3189@gmail.com>                             //
 //   http://www.icplusplus.com                                           //
@@ -56,8 +56,12 @@
 #define _STATUS_TO_FILE 1
 
 #define _ENABLE_VERBOSE_LOGS 1 
-#define _VERBOSE_TO_STANDARD_OUTPUT 1
-#define _VERBOSE_TO_FILE 1
+#define _VERBOSE_LOGS_TO_STANDARD_OUTPUT 1
+#define _VERBOSE_LOGS_TO_FILE 1
+
+#define _ENABLE_QA_LOGS 1
+#define _QA_LOGS_TO_STANDARD_OUTPUT 1
+#define _QA_LOGS_TO_FILE 1
 
 ////////////////////////////////////////////////////////////////////
 ///                                                              ///
@@ -107,13 +111,13 @@
  #include <sys/stat.h>
  #include <sys/time.h>
 #endif //_LINUX || _MAC
+#include <string>
 #include <list>
 #include <iostream>
 #include <sstream>
 #include <fstream>
 
 namespace easyloggingpp {
-
 ///////////////////////////////////////////////////////////////////////
 ///                                                                 ///
 ///                   CONFIGURATION FOR LOGGING                     ///
@@ -121,7 +125,7 @@ namespace easyloggingpp {
 ///                       MODIFIABLE SECTION                        ///
 ///                                                                 ///
 ///////////////////////////////////////////////////////////////////////
-
+namespace configuration {
 
 // FOR DETAILS ON FOLLOWING CONFIGURATION PLEASE SEE README AT:
 // https://github.com/mkhan3189/EasyLoggingPP/blob/master/README.md
@@ -136,6 +140,7 @@ const std::string    PERFORMANCE_LOG_FORMAT   =    DEFAULT_LOG_FORMAT;
 const std::string    HINT_LOG_FORMAT          =    DEFAULT_LOG_FORMAT;
 const std::string    STATUS_LOG_FORMAT        =    DEFAULT_LOG_FORMAT;
 const std::string    VERBOSE_LOG_FORMAT       =    "[%level-%vlevel] [%datetime] %log\n";
+const std::string    QA_LOG_FORMAT            =    DEFAULT_LOG_FORMAT;
 
 const bool           SHOW_STD_OUTPUT          =    true;
 const bool           SAVE_TO_FILE             =    true;
@@ -144,6 +149,7 @@ const bool           USE_CUSTOM_LOCATION      =    true;
 const std::string    CUSTOM_LOG_FILE_LOCATION =    "logs/";
 const bool           SHOW_START_FUNCTION_LOG  =    false;
 
+} // namespace configuration
 ////////////////////////////////////////////////////////////////////
 ///                                                              ///
 ///                    END OF CONFIGURATION                      ///
@@ -181,19 +187,21 @@ const bool           SHOW_START_FUNCTION_LOG  =    false;
 #define _HINT_LOG        ((_ENABLE_HINT) && !defined(_DISABLE_HINT))
 #define _FATAL_LOG       ((_ENABLE_FATAL_LOGS) && !defined(_DISABLE_FATAL_LOGS))
 #define _VERBOSE_LOG     ((_ENABLE_VERBOSE_LOGS) && !defined(_DISABLE_VERBOSE_LOGS))
+#define _QA_LOG          ((_ENABLE_QA_LOGS) && defined(_QUALITY_ASSURANCE))
 
 // Application arguments based logs
-#define _START_EASYLOGGINGPP(argc, argv) ::easyloggingpp::setAppArgs(argc, argv);
-#define _END_EASYLOGGINGPP ::easyloggingpp::releaseMemory();
+#define _START_EASYLOGGINGPP(argc, argv) ::easyloggingpp::internal::setAppArgs(argc, argv);
+#define _END_EASYLOGGINGPP ::easyloggingpp::internal::releaseMemory();
 
+namespace internal {
 //
 // Static fields
 //
-static const std::string kFinalFilename = (::easyloggingpp::USE_CUSTOM_LOCATION ? ::easyloggingpp::CUSTOM_LOG_FILE_LOCATION : "") + ::easyloggingpp::LOG_FILENAME;
-static bool showDateTime = ::easyloggingpp::DEFAULT_LOG_FORMAT.find("%datetime") != std::string::npos;
-static bool showDate = (!::easyloggingpp::showDateTime) && (::easyloggingpp::DEFAULT_LOG_FORMAT.find("%date") != std::string::npos);
-static bool showTime = (!::easyloggingpp::showDateTime) && (::easyloggingpp::DEFAULT_LOG_FORMAT.find("%time") != std::string::npos);
-static bool showLocation = ::easyloggingpp::DEFAULT_LOG_FORMAT.find("%loc") != std::string::npos;
+static const std::string kFinalFilename = (::easyloggingpp::configuration::USE_CUSTOM_LOCATION ? ::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION : "") + ::easyloggingpp::configuration::LOG_FILENAME;
+static bool showDateTime = ::easyloggingpp::configuration::DEFAULT_LOG_FORMAT.find("%datetime") != std::string::npos;
+static bool showDate = (!::easyloggingpp::internal::showDateTime) && (::easyloggingpp::configuration::DEFAULT_LOG_FORMAT.find("%date") != std::string::npos);
+static bool showTime = (!::easyloggingpp::internal::showDateTime) && (::easyloggingpp::configuration::DEFAULT_LOG_FORMAT.find("%time") != std::string::npos);
+static bool showLocation = ::easyloggingpp::configuration::DEFAULT_LOG_FORMAT.find("%loc") != std::string::npos;
 static std::string user;
 static std::string host;
 static std::stringstream *logStream = NULL;
@@ -226,7 +234,7 @@ public:
   bool toStandardOutput;
   bool toFile;
 };
-static std::list< ::easyloggingpp::LogType > logTypes;
+static std::list< ::easyloggingpp::internal::LogType > logTypes;
 
 //
 // Internal functions
@@ -243,7 +251,7 @@ static inline void setAppArgs(int argc, char** argv) {
      && (argv[argc][1] == '-')
      && (argv[argc][2] == 'v')
      && (argv[argc][3] == '=')) {
-       ::easyloggingpp::verboseLevel = atoi(argv[argc] + 4);
+       ::easyloggingpp::internal::verboseLevel = atoi(argv[argc] + 4);
     }
 #endif //_VERBOSE_LOG
   }
@@ -251,29 +259,29 @@ static inline void setAppArgs(int argc, char** argv) {
 
 static bool logPathExist(void) {
 #if _WINDOWS
-  DWORD fileType = GetFileAttributesA(::easyloggingpp::CUSTOM_LOG_FILE_LOCATION.c_str());
+  DWORD fileType = GetFileAttributesA(::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION.c_str());
   if (fileType == INVALID_FILE_ATTRIBUTES) {
     return false;
   }
   return (fileType & FILE_ATTRIBUTE_DIRECTORY);
 #elif _LINUX || _MAC
   struct stat st;
-  return (stat(::easyloggingpp::CUSTOM_LOG_FILE_LOCATION.c_str(), &st) == 0);
+  return (stat(::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION.c_str(), &st) == 0);
 #endif //_WINDOWS
 }
 
 static void createLogPath(void) {
 #if _WINDOWS || _LINUX || _MAC
-  if ((::easyloggingpp::USE_CUSTOM_LOCATION)
-   && (::easyloggingpp::CUSTOM_LOG_FILE_LOCATION.size() > 0)
-   && (!::easyloggingpp::logPathExist())) {
+  if ((::easyloggingpp::configuration::USE_CUSTOM_LOCATION)
+   && (::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION.size() > 0)
+   && (!::easyloggingpp::internal::logPathExist())) {
     int status = -1;
  #if _WINDOWS
     std::string pathDelimiter = "\\";
  #elif _LINUX || _MAC
     std::string pathDelimiter = "/";
  #endif //_WINDOWS
-    std::string tempPath = CUSTOM_LOG_FILE_LOCATION;
+    std::string tempPath = ::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION;
     std::string dir = "";
     short foundAt = -1;
     std::string madeSoFar = tempPath.substr(0, 1) == pathDelimiter ? pathDelimiter : "";
@@ -290,16 +298,16 @@ static void createLogPath(void) {
       tempPath = tempPath.erase(0, foundAt + 1);
     }
     if (status == -1) {
-      ::easyloggingpp::internalMessage("Unable to create log path [" + ::easyloggingpp::CUSTOM_LOG_FILE_LOCATION + "]");
+      ::easyloggingpp::internal::internalMessage("Unable to create log path [" + ::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION + "]");
     }
   }
 #else
-  ::easyloggingpp::internalMessage("Unable to create log path [" + ::easyloggingpp::CUSTOM_LOG_FILE_LOCATION + "]");
+  ::easyloggingpp::internal::internalMessage("Unable to create log path [" + ::easyloggingpp::configuration::CUSTOM_LOG_FILE_LOCATION + "]");
 #endif //_WINDOWS || _LINUX || _MAC
 }
 
 static inline std::string getDateTime(void) {
-  if (!(::easyloggingpp::showDateTime || ::easyloggingpp::showDate || ::easyloggingpp::showTime)) return "";
+  if (!(::easyloggingpp::internal::showDateTime || ::easyloggingpp::internal::showDate || ::easyloggingpp::internal::showTime)) return "";
 #if _WINDOWS
     time_t currTime;
     time(&currTime);
@@ -307,7 +315,7 @@ static inline std::string getDateTime(void) {
     timeval currTime;
     gettimeofday(&currTime, NULL);
     int milliSeconds = 0;
-    if ((::easyloggingpp::showDateTime) || (::easyloggingpp::showTime)) {
+    if ((::easyloggingpp::internal::showDateTime) || (::easyloggingpp::internal::showTime)) {
       milliSeconds = currTime.tv_usec / 1000;
     }
 #endif //_WINDOWS
@@ -317,13 +325,13 @@ static inline std::string getDateTime(void) {
 #elif _LINUX || _MAC
     timeInfo = localtime(&currTime.tv_sec);
 #endif //_WINDOWS
-    strftime(::easyloggingpp::dateBuffer, ::easyloggingpp::kDateBufferSize, ::easyloggingpp::dateFormat, timeInfo);
+    strftime(::easyloggingpp::internal::dateBuffer, ::easyloggingpp::internal::kDateBufferSize, ::easyloggingpp::internal::dateFormat, timeInfo);
 #if _LINUX || _MAC
-    if ((::easyloggingpp::showDateTime) || (::easyloggingpp::showTime)) {
-      sprintf(::easyloggingpp::dateBuffer, "%s.%d", ::easyloggingpp::dateBuffer, milliSeconds);
+    if ((::easyloggingpp::internal::showDateTime) || (::easyloggingpp::internal::showTime)) {
+      sprintf(::easyloggingpp::internal::dateBuffer, "%s.%d", ::easyloggingpp::internal::dateBuffer, milliSeconds);
     }
 #endif //_LINUX || _MAC
-  return std::string(::easyloggingpp::dateBuffer);
+  return std::string(::easyloggingpp::internal::dateBuffer);
 }
 
 static inline std::string getUsername(void) {
@@ -353,125 +361,108 @@ static inline std::string getHostname(void) {
 }
 
 static inline void cleanStream(void) {
-  ::easyloggingpp::tempStream.str("");
-  ::easyloggingpp::tempStream2.str("");
-  ::easyloggingpp::logStream->str("");
+  ::easyloggingpp::internal::tempStream.str("");
+  ::easyloggingpp::internal::tempStream2.str("");
+  ::easyloggingpp::internal::logStream->str("");
 }
 
 static inline void releaseMemory(void) {
-  ::easyloggingpp::cleanStream();
-  delete ::easyloggingpp::logFile;
-  delete ::easyloggingpp::logStream;
-  ::easyloggingpp::loggerInitialized = false;
+  ::easyloggingpp::internal::cleanStream();
+  delete ::easyloggingpp::internal::logFile;
+  delete ::easyloggingpp::internal::logStream;
+  ::easyloggingpp::internal::loggerInitialized = false;
 }
 
 static inline void updateDateFormat(void) {
   const char* dateFormatLocal = "%d/%m/%Y";
   const char* timeFormatLocal = "%H:%M:%S";
-  if (::easyloggingpp::showDate) {
-    strcpy(::easyloggingpp::dateFormat, dateFormatLocal);
-  } else if (::easyloggingpp::showTime) {
-    strcpy(::easyloggingpp::dateFormat, timeFormatLocal);
-  } else if (::easyloggingpp::showDateTime) {
-    strcpy(::easyloggingpp::dateFormat, dateFormatLocal);
-    strcat(::easyloggingpp::dateFormat, " ");
-    strcat(::easyloggingpp::dateFormat, timeFormatLocal);
+  if (::easyloggingpp::internal::showDate) {
+    strcpy(::easyloggingpp::internal::dateFormat, dateFormatLocal);
+  } else if (::easyloggingpp::internal::showTime) {
+    strcpy(::easyloggingpp::internal::dateFormat, timeFormatLocal);
+  } else if (::easyloggingpp::internal::showDateTime) {
+    strcpy(::easyloggingpp::internal::dateFormat, dateFormatLocal);
+    strcat(::easyloggingpp::internal::dateFormat, " ");
+    strcat(::easyloggingpp::internal::dateFormat, timeFormatLocal);
   }
 }
 
 static void init(void) {
-  if (!::easyloggingpp::loggerInitialized) {
+  if (!::easyloggingpp::internal::loggerInitialized) {
     // Logger
-    if (::easyloggingpp::SHOW_STD_OUTPUT) {
-      ::easyloggingpp::logStream = new std::stringstream();
+    if (::easyloggingpp::configuration::SHOW_STD_OUTPUT) {
+      ::easyloggingpp::internal::logStream = new std::stringstream();
     }
     // Path
-    ::easyloggingpp::createLogPath();
+    ::easyloggingpp::internal::createLogPath();
     // Log file
-    if (::easyloggingpp::SAVE_TO_FILE) {
+    if (::easyloggingpp::configuration::SAVE_TO_FILE) {
 #if defined(_ALWAYS_CLEAN_LOGS)
         std::ios_base::openmode mode = std::ofstream::out;
 #else
         std::ios_base::openmode mode = std::ofstream::out | std::ofstream::app;
 #endif //defined(_ALWAYS_CLEAN_LOGS)
-      ::easyloggingpp::logFile = new std::ofstream(::easyloggingpp::kFinalFilename.c_str(), mode);
-      if ((!::easyloggingpp::fileNotOpenedErrorDisplayed) && (!::easyloggingpp::logFile->is_open())) {
-        ::easyloggingpp::internalMessage("Unable to open log file [" + ::easyloggingpp::kFinalFilename + "]");
-        ::easyloggingpp::fileNotOpenedErrorDisplayed = true;
+      ::easyloggingpp::internal::logFile = new std::ofstream(::easyloggingpp::internal::kFinalFilename.c_str(), mode);
+      if ((!::easyloggingpp::internal::fileNotOpenedErrorDisplayed) && (!::easyloggingpp::internal::logFile->is_open())) {
+        ::easyloggingpp::internal::internalMessage("Unable to open log file [" + ::easyloggingpp::internal::kFinalFilename + "]");
+        ::easyloggingpp::internal::fileNotOpenedErrorDisplayed = true;
       } else {
-        ::easyloggingpp::logFile->close();
+        ::easyloggingpp::internal::logFile->close();
       }
     }
     // Date format
-    ::easyloggingpp::updateDateFormat();
+    ::easyloggingpp::internal::updateDateFormat();
     // Username and host
-    ::easyloggingpp::user = ::easyloggingpp::getUsername();
-    ::easyloggingpp::host = ::easyloggingpp::getHostname();
+    ::easyloggingpp::internal::user = ::easyloggingpp::internal::getUsername();
+    ::easyloggingpp::internal::host = ::easyloggingpp::internal::getHostname();
     // Different log levels
 #if _DEBUG_LOG
-     logTypes.push_back(LogType("DEBUG", ::easyloggingpp::DEBUG_LOG_FORMAT, _DEBUG_LOGS_TO_STANDARD_OUTPUT, _DEBUG_LOGS_TO_FILE));
+     logTypes.push_back(LogType("DEBUG", ::easyloggingpp::configuration::DEBUG_LOG_FORMAT, _DEBUG_LOGS_TO_STANDARD_OUTPUT, _DEBUG_LOGS_TO_FILE));
 #endif //_DEBUG_LOG
 #if _INFO_LOG
-     logTypes.push_back(LogType("INFO", ::easyloggingpp::INFO_LOG_FORMAT, _INFO_LOGS_TO_STANDARD_OUTPUT, _INFO_LOGS_TO_FILE));
+     logTypes.push_back(LogType("INFO", ::easyloggingpp::configuration::INFO_LOG_FORMAT, _INFO_LOGS_TO_STANDARD_OUTPUT, _INFO_LOGS_TO_FILE));
 #endif //_INFO_LOG
 #if _WARNING_LOG
-     logTypes.push_back(LogType("WARNING", ::easyloggingpp::WARNING_LOG_FORMAT, _WARNING_LOGS_TO_STANDARD_OUTPUT, _WARNING_LOGS_TO_FILE));
+     logTypes.push_back(LogType("WARNING", ::easyloggingpp::configuration::WARNING_LOG_FORMAT, _WARNING_LOGS_TO_STANDARD_OUTPUT, _WARNING_LOGS_TO_FILE));
 #endif //_WARNING_LOG
 #if _ERROR_LOG
-     logTypes.push_back(LogType("ERROR", ::easyloggingpp::ERROR_LOG_FORMAT, _ERROR_LOGS_TO_STANDARD_OUTPUT, _ERROR_LOGS_TO_FILE));
+     logTypes.push_back(LogType("ERROR", ::easyloggingpp::configuration::ERROR_LOG_FORMAT, _ERROR_LOGS_TO_STANDARD_OUTPUT, _ERROR_LOGS_TO_FILE));
 #endif //_ERROR_LOG
 #if _FATAL_LOG
-     logTypes.push_back(LogType("FATAL", ::easyloggingpp::FATAL_LOG_FORMAT, _FATAL_LOGS_TO_STANDARD_OUTPUT, _FATAL_LOGS_TO_FILE));
+     logTypes.push_back(LogType("FATAL", ::easyloggingpp::configuration::FATAL_LOG_FORMAT, _FATAL_LOGS_TO_STANDARD_OUTPUT, _FATAL_LOGS_TO_FILE));
 #endif //_FATAL_LOG
 #if _PERFORMANCE_LOG
-     logTypes.push_back(LogType("PERFORMANCE", ::easyloggingpp::PERFORMANCE_LOG_FORMAT, _PERFORMANCE_LOGS_TO_STANDARD_OUTPUT, _PERFORMANCE_LOGS_TO_FILE));
+     logTypes.push_back(LogType("PERFORMANCE", ::easyloggingpp::configuration::PERFORMANCE_LOG_FORMAT, _PERFORMANCE_LOGS_TO_STANDARD_OUTPUT, _PERFORMANCE_LOGS_TO_FILE));
 #endif //_PERFORMANCE_LOG
 #if _HINT_LOG
-     logTypes.push_back(LogType("HINT", ::easyloggingpp::HINT_LOG_FORMAT, _HINT_TO_STANDARD_OUTPUT, _HINT_TO_FILE));
+     logTypes.push_back(LogType("HINT", ::easyloggingpp::configuration::HINT_LOG_FORMAT, _HINT_TO_STANDARD_OUTPUT, _HINT_TO_FILE));
 #endif //_HINT_LOG
 #if _STATUS_LOG
-     logTypes.push_back(LogType("STATUS", ::easyloggingpp::STATUS_LOG_FORMAT, _STATUS_TO_STANDARD_OUTPUT, _STATUS_TO_FILE));
+     logTypes.push_back(LogType("STATUS", ::easyloggingpp::configuration::STATUS_LOG_FORMAT, _STATUS_TO_STANDARD_OUTPUT, _STATUS_TO_FILE));
 #endif //_STATUS_LOG
 #if _VERBOSE_LOG
-     logTypes.push_back(LogType("VERBOSE", ::easyloggingpp::VERBOSE_LOG_FORMAT, _VERBOSE_TO_STANDARD_OUTPUT, _VERBOSE_TO_FILE));
+     logTypes.push_back(LogType("VERBOSE", ::easyloggingpp::configuration::VERBOSE_LOG_FORMAT, _VERBOSE_LOGS_TO_STANDARD_OUTPUT, _VERBOSE_LOGS_TO_FILE));
 #endif //_VERBOSE_LOG
-     ::easyloggingpp::loggerInitialized = true;
+#if _QA_LOG
+     logTypes.push_back(LogType("QA", ::easyloggingpp::configuration::QA_LOG_FORMAT, _QA_LOGS_TO_STANDARD_OUTPUT, _QA_LOGS_TO_FILE));
+#endif //_QA_LOG
+     ::easyloggingpp::internal::loggerInitialized = true;
    }
 }
 
-static std::string readLog(void) {
-  std::stringstream ss;
-  if (::easyloggingpp::SAVE_TO_FILE) {
-    std::ifstream logFileReader(::easyloggingpp::kFinalFilename.c_str(), std::ifstream::in);
-    if (logFileReader.is_open()) {
-      std::string line;
-      while (logFileReader.good()) {
-        std::getline(logFileReader, line);
-        ss << line << std::endl;
-      }
-      logFileReader.close();
-    } else {
-      ss << "Error opening log file [" << ::easyloggingpp::kFinalFilename << "]";
-    }
-  } else {
-    ss << "Logs are not being saved to file!";
-  }
-  return ss.str();
-}
-
 static inline void writeLog(void) {
-  if ((::easyloggingpp::logStream) && (::easyloggingpp::toStandardOutput)) {
-    std::cout << ::easyloggingpp::logStream->str();
+  if ((::easyloggingpp::internal::logStream) && (::easyloggingpp::internal::toStandardOutput)) {
+    std::cout << ::easyloggingpp::internal::logStream->str();
   }
-  if ((!::easyloggingpp::fileNotOpenedErrorDisplayed)
-   && (::easyloggingpp::logStream)
-   && (::easyloggingpp::logFile)
-   && (::easyloggingpp::toFile)) {
-    ::easyloggingpp::logFile->open(::easyloggingpp::kFinalFilename.c_str(), std::ofstream::out | std::ofstream::app);
-    (*::easyloggingpp::logFile) << ::easyloggingpp::logStream->str();
-    ::easyloggingpp::logFile->close();
+  if ((!::easyloggingpp::internal::fileNotOpenedErrorDisplayed)
+   && (::easyloggingpp::internal::logStream)
+   && (::easyloggingpp::internal::logFile)
+   && (::easyloggingpp::internal::toFile)) {
+    ::easyloggingpp::internal::logFile->open(::easyloggingpp::internal::kFinalFilename.c_str(), std::ofstream::out | std::ofstream::app);
+    (*::easyloggingpp::internal::logFile) << ::easyloggingpp::internal::logStream->str();
+    ::easyloggingpp::internal::logFile->close();
   }
-  ::easyloggingpp::cleanStream();
+  ::easyloggingpp::internal::cleanStream();
 }
 
 static void updateFormatValue(const std::string& formatSpecifier, const std::string& value, std::string& currentFormat) {
@@ -488,71 +479,71 @@ static void updateFormatValue(const std::string& formatSpecifier, const std::str
 }
 
 static void determineCommonLogFormat(const std::string& format) {
-  ::easyloggingpp::logFormat = format;
-  ::easyloggingpp::showDateTime = format.find("%datetime") != std::string::npos;
-  if (!::easyloggingpp::showDateTime) {
-    ::easyloggingpp::showDate = (format.find("%date") != std::string::npos);
-    ::easyloggingpp::showTime = (format.find("%time") != std::string::npos);
+  ::easyloggingpp::internal::logFormat = format;
+  ::easyloggingpp::internal::showDateTime = format.find("%datetime") != std::string::npos;
+  if (!::easyloggingpp::internal::showDateTime) {
+    ::easyloggingpp::internal::showDate = (format.find("%date") != std::string::npos);
+    ::easyloggingpp::internal::showTime = (format.find("%time") != std::string::npos);
   }
-  ::easyloggingpp::showLocation = format.find("%loc") != std::string::npos;
-  ::easyloggingpp::updateDateFormat();
+  ::easyloggingpp::internal::showLocation = format.find("%loc") != std::string::npos;
+  ::easyloggingpp::internal::updateDateFormat();
 }
 
 static void determineLogFormat(const std::string& type) {
-  for (std::list< ::easyloggingpp::LogType >::iterator it = logTypes.begin();
+  for (std::list< ::easyloggingpp::internal::LogType >::iterator it = logTypes.begin();
        it != logTypes.end();
        ++it) {
     if (type == it->name) {
-      ::easyloggingpp::determineCommonLogFormat(it->format);
-      ::easyloggingpp::toStandardOutput = it->toStandardOutput;
-      ::easyloggingpp::toFile = it->toFile;
+      ::easyloggingpp::internal::determineCommonLogFormat(it->format);
+      ::easyloggingpp::internal::toStandardOutput = it->toStandardOutput;
+      ::easyloggingpp::internal::toFile = it->toFile;
       return;
     }
   }
 }
 
 static void buildFormat(const char* func, const char* file, const unsigned long int line, const std::string& type, int verboseLevel = -1) {
-  ::easyloggingpp::init();
-  ::easyloggingpp::determineLogFormat(type);
-  ::easyloggingpp::updateFormatValue("%level", type, ::easyloggingpp::logFormat);
+  ::easyloggingpp::internal::init();
+  ::easyloggingpp::internal::determineLogFormat(type);
+  ::easyloggingpp::internal::updateFormatValue("%level", type, ::easyloggingpp::internal::logFormat);
 #if _VERBOSE_LOG
   if (verboseLevel != -1) {
-    ::easyloggingpp::tempStream << verboseLevel;
-    ::easyloggingpp::updateFormatValue("%vlevel", ::easyloggingpp::tempStream.str(), ::easyloggingpp::logFormat);
-    ::easyloggingpp::tempStream.str("");
+    ::easyloggingpp::internal::tempStream << verboseLevel;
+    ::easyloggingpp::internal::updateFormatValue("%vlevel", ::easyloggingpp::internal::tempStream.str(), ::easyloggingpp::internal::logFormat);
+    ::easyloggingpp::internal::tempStream.str("");
   }
 #endif //_VERBOSE_LOG
-  if (::easyloggingpp::showDateTime) {
-    ::easyloggingpp::updateFormatValue("%datetime", ::easyloggingpp::getDateTime(), ::easyloggingpp::logFormat);
-  } else if (::easyloggingpp::showDate) {
-    ::easyloggingpp::updateFormatValue("%date", ::easyloggingpp::getDateTime(), ::easyloggingpp::logFormat);
-  } else if (::easyloggingpp::showTime) {
-    ::easyloggingpp::updateFormatValue("%time", ::easyloggingpp::getDateTime(), ::easyloggingpp::logFormat);
+  if (::easyloggingpp::internal::showDateTime) {
+    ::easyloggingpp::internal::updateFormatValue("%datetime", ::easyloggingpp::internal::getDateTime(), ::easyloggingpp::internal::logFormat);
+  } else if (::easyloggingpp::internal::showDate) {
+    ::easyloggingpp::internal::updateFormatValue("%date", ::easyloggingpp::internal::getDateTime(), ::easyloggingpp::internal::logFormat);
+  } else if (::easyloggingpp::internal::showTime) {
+    ::easyloggingpp::internal::updateFormatValue("%time", ::easyloggingpp::internal::getDateTime(), ::easyloggingpp::internal::logFormat);
   }
-  ::easyloggingpp::updateFormatValue("%func", std::string(func), ::easyloggingpp::logFormat);
-  if (::easyloggingpp::showLocation) {
-    ::easyloggingpp::tempStream << file << ":" << line;
-    ::easyloggingpp::updateFormatValue("%loc", ::easyloggingpp::tempStream.str(), ::easyloggingpp::logFormat);
+  ::easyloggingpp::internal::updateFormatValue("%func", std::string(func), ::easyloggingpp::internal::logFormat);
+  if (::easyloggingpp::internal::showLocation) {
+    ::easyloggingpp::internal::tempStream << file << ":" << line;
+    ::easyloggingpp::internal::updateFormatValue("%loc", ::easyloggingpp::internal::tempStream.str(), ::easyloggingpp::internal::logFormat);
   }
-  ::easyloggingpp::updateFormatValue("%user", ::easyloggingpp::user, ::easyloggingpp::logFormat);
-  ::easyloggingpp::updateFormatValue("%host", ::easyloggingpp::host, ::easyloggingpp::logFormat);
-  ::easyloggingpp::updateFormatValue("%log", ::easyloggingpp::tempStream2.str(), ::easyloggingpp::logFormat);
-  (*::easyloggingpp::logStream) << logFormat;
+  ::easyloggingpp::internal::updateFormatValue("%user", ::easyloggingpp::internal::user, ::easyloggingpp::internal::logFormat);
+  ::easyloggingpp::internal::updateFormatValue("%host", ::easyloggingpp::internal::host, ::easyloggingpp::internal::logFormat);
+  ::easyloggingpp::internal::updateFormatValue("%log", ::easyloggingpp::internal::tempStream2.str(), ::easyloggingpp::internal::logFormat);
+  (*::easyloggingpp::internal::logStream) << logFormat;
 }
 
 //
 // Logging macros
 //
 #define WRITE_LOG(type, log)\
-  ::easyloggingpp::tempStream2 << log;\
-  ::easyloggingpp::buildFormat(__func__, __FILE__, __LINE__, std::string(type));\
-  ::easyloggingpp::writeLog();
+  ::easyloggingpp::internal::tempStream2 << log;\
+  ::easyloggingpp::internal::buildFormat(__func__, __FILE__, __LINE__, std::string(type));\
+  ::easyloggingpp::internal::writeLog();
 
 #define WRITE_VLOG(level, log)\
-  if (::easyloggingpp::verboseLevel >= level) {\
-    ::easyloggingpp::tempStream2 << log;\
-    ::easyloggingpp::buildFormat(__func__, __FILE__, __LINE__, std::string("VERBOSE"), level);\
-    ::easyloggingpp::writeLog();\
+  if (::easyloggingpp::internal::verboseLevel >= level) {\
+    ::easyloggingpp::internal::tempStream2 << log;\
+    ::easyloggingpp::internal::buildFormat(__func__, __FILE__, __LINE__, std::string("VERBOSE"), level);\
+    ::easyloggingpp::internal::writeLog();\
   }
 
   #if _DEBUG_LOG
@@ -611,8 +602,8 @@ static void buildFormat(const char* func, const char* file, const unsigned long 
     #define PERFORMANCE(logMessage) WRITE_LOG("PERFORMANCE",logMessage)
     #define PERFORMANCE_IF(condition, logMessage) if (condition) { PERFORMANCE(logMessage); }
     #define START_FUNCTION_LOG "Executing [" << __func__ << "]"
-    #define TIME_OUTPUT "Executed [" << __func__ << "] in [~ " << ::easyloggingpp::formatSeconds(difftime(functionEndTime,functionStartTime)) << "]"
-    #define FUNC_SUB_COMMON_START { if (::easyloggingpp::SHOW_START_FUNCTION_LOG) { PERFORMANCE(START_FUNCTION_LOG) } time_t functionStartTime, functionEndTime; time(&functionStartTime);
+    #define TIME_OUTPUT "Executed [" << __func__ << "] in [~ " << ::easyloggingpp::internal::formatSeconds(difftime(functionEndTime,functionStartTime)) << "]"
+    #define FUNC_SUB_COMMON_START { if (::easyloggingpp::configuration::SHOW_START_FUNCTION_LOG) { PERFORMANCE(START_FUNCTION_LOG) } time_t functionStartTime, functionEndTime; time(&functionStartTime);
     #define FUNC_SUB_COMMON_END time(&functionEndTime); PERFORMANCE(TIME_OUTPUT);
     #define SUB(FUNCTION_NAME,PARAMS) void FUNCTION_NAME PARAMS FUNC_SUB_COMMON_START
     #define END_SUB FUNC_SUB_COMMON_END }
@@ -652,7 +643,37 @@ static void buildFormat(const char* func, const char* file, const unsigned long 
     #define VERBOSE(x, y)
     #define VERBOSE_IF(x, y, z)
   #endif //_VERBOSE_LOG
-} //namespace easyloggingpp
+
+  #if _QA_LOG
+    #define QA(logMessage) WRITE_LOG("QA",logMessage)
+    #define QA_IF(condition, logMessage) if (condition) { QA(logMessage); }
+  #else
+    #define QA(x)
+    #define QA_IF(x, y)
+  #endif //_QA_LOG
+} // namespace internal
+namespace helper {
+  static std::string readLog(void) {
+   std::stringstream ss;
+   if (::easyloggingpp::configuration::SAVE_TO_FILE) {
+     std::ifstream logFileReader(::easyloggingpp::internal::kFinalFilename.c_str(), std::ifstream::in);
+     if (logFileReader.is_open()) {
+       std::string line;
+       while (logFileReader.good()) {
+         std::getline(logFileReader, line);
+         ss << line << std::endl;
+       }
+       logFileReader.close();
+     } else {
+       ss << "Error opening log file [" << ::easyloggingpp::internal::kFinalFilename << "]";
+     }
+   } else {
+     ss << "Logs are not being saved to file!";
+   }
+  return ss.str();
+  }
+} // namespace helper
+} // namespace easyloggingpp
 #else
   // Essentials
   #define DEBUG(x)
@@ -664,6 +685,7 @@ static void buildFormat(const char* func, const char* file, const unsigned long 
   #define HINT(x)
   #define STATUS(x)
   #define VERBOSE(x, y)
+  #define QA(x)
   // Performance specific
   #define SUB(FUNCTION_NAME,PARAMS) void FUNCTION_NAME PARAMS {
   #define END_SUB }
@@ -680,6 +702,18 @@ static void buildFormat(const char* func, const char* file, const unsigned long 
   #define HINT_IF(x, y)
   #define STATUS_IF(x, y)
   #define VERBOSE_IF(x, y, z)
-#endif //(defined(_LOGGING_ENABLED) && !defined(_DISABLE_LOGS))
+  #define QA_IF(x, y)
+  // Miscellaneous
+  #define _START_EASYLOGGINGPP(x, y)
+  #define _END_EASYLOGGINGPP
+  // Helper functions
+  #include <string>
+  namespace easyloggingpp {
+    namespace helper {
+      std::string readLog() {
+        return "";
+      }
+    } // namespace helper
+  } // namespace easyloggingpp
+#endif //((_LOGGING_ENABLED) && !defined(_DISABLE_LOGS))
 #endif //EASYLOGGINGPP_H
-